\name{rngex-package}
\alias{rngex-package}
\alias{rngex}
\docType{package}
\title{
\packageTitle{rngex}
}
\description{
\packageDescription{rngex}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{rngex}
\packageIndices{rngex}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{rngex}

Maintainer: \packageMaintainer{rngex}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
